Rename Integrator and SteadyFlightPhase
