+++ openconcept/analysis/performance/mission_profiles.py
@@ -1,6 +1,6 @@
 from openmdao.api import IndepVarComp, Group, BalanceComp, ExecComp
 import openconcept.api as oc
-from openconcept.analysis.performance.solver_phases import BFLImplicitSolve, GroundRollPhase, RotationPhase, RobustRotationPhase, SteadyFlightPhase, ClimbAnglePhase, NewSteadyFlightPhase
+from openconcept.analysis.performance.solver_phases import BFLImplicitSolve, GroundRollPhase, RotationPhase, RobustRotationPhase, ClimbAnglePhase, SteadyFlightPhase
 from openconcept.utilities.dvlabel import DVLabel
 
 # class ThreePhaseMissionOnly(Group):
@@ -166,17 +166,17 @@ class MissionWithReserve(oc.TrajectoryGroup):
             mp.add_output('payload',val=1000.,units='lbm')
 
             # add the climb, cruise, and descent segments
-            phase1 = self.add_subsystem('climb',NewSteadyFlightPhase(num_nodes=nn, aircraft_model=acmodelclass, flight_phase='climb'),promotes_inputs=['ac|*'])
+            phase1 = self.add_subsystem('climb',SteadyFlightPhase(num_nodes=nn, aircraft_model=acmodelclass, flight_phase='climb'),promotes_inputs=['ac|*'])
             # set the climb time such that the specified initial cruise altitude is exactly reached
             phase1.add_subsystem('climbdt',BalanceComp(name='duration',units='s',eq_units='m',val=120,upper=2000,lower=0,rhs_name='cruise|h0',lhs_name='fltcond|h_final'),promotes_outputs=['duration'])
             phase1.connect('ode_integ.fltcond|h_final','climbdt.fltcond|h_final')
             self.connect('cruise|h0', 'climb.climbdt.cruise|h0')
 
-            phase2 = self.add_subsystem('cruise',NewSteadyFlightPhase(num_nodes=nn, aircraft_model=acmodelclass, flight_phase='cruise'),promotes_inputs=['ac|*'])
+            phase2 = self.add_subsystem('cruise',SteadyFlightPhase(num_nodes=nn, aircraft_model=acmodelclass, flight_phase='cruise'),promotes_inputs=['ac|*'])
             # set the cruise time such that the desired design range is flown by the end of the mission
             phase2.add_subsystem('cruisedt',BalanceComp(name='duration',units='s',eq_units='m',val=120, upper=25000, lower=0,rhs_name='mission_range',lhs_name='range_final'),promotes_outputs=['duration'])
             self.connect('mission_range', 'cruise.cruisedt.mission_range')
-            phase3 = self.add_subsystem('descent',NewSteadyFlightPhase(num_nodes=nn, aircraft_model=acmodelclass, flight_phase='descent'),promotes_inputs=['ac|*'])
+            phase3 = self.add_subsystem('descent',SteadyFlightPhase(num_nodes=nn, aircraft_model=acmodelclass, flight_phase='descent'),promotes_inputs=['ac|*'])
             # set the descent time so that the final altitude is sea level again
             phase3.add_subsystem('descentdt',BalanceComp(name='duration',units='s',eq_units='m', val=120, upper=8000, lower=0,rhs_name='takeoff|h',lhs_name='fltcond|h_final'),promotes_outputs=['duration'])
             self.connect('descent.ode_integ.range_final','cruise.cruisedt.range_final')
@@ -184,18 +184,18 @@ class MissionWithReserve(oc.TrajectoryGroup):
             phase3.connect('ode_integ.fltcond|h_final','descentdt.fltcond|h_final')
 
             # add the climb, cruise, and descent segments for the reserve mission
-            phase4 = self.add_subsystem('reserve_climb',NewSteadyFlightPhase(num_nodes=nn, aircraft_model=acmodelclass, flight_phase='reserve_climb'),promotes_inputs=['ac|*'])
+            phase4 = self.add_subsystem('reserve_climb',SteadyFlightPhase(num_nodes=nn, aircraft_model=acmodelclass, flight_phase='reserve_climb'),promotes_inputs=['ac|*'])
             # set the climb time such that the specified initial cruise altitude is exactly reached
             phase4.add_subsystem('reserve_climbdt',BalanceComp(name='duration',units='s',eq_units='m',val=120,upper=2000,lower=0,rhs_name='reserve|h0',lhs_name='fltcond|h_final'),promotes_outputs=['duration'])
             phase4.connect('ode_integ.fltcond|h_final','reserve_climbdt.fltcond|h_final')
             self.connect('reserve|h0', 'reserve_climb.reserve_climbdt.reserve|h0')
 
-            phase5 = self.add_subsystem('reserve_cruise',NewSteadyFlightPhase(num_nodes=nn, aircraft_model=acmodelclass, flight_phase='reserve_cruise'),promotes_inputs=['ac|*'])
+            phase5 = self.add_subsystem('reserve_cruise',SteadyFlightPhase(num_nodes=nn, aircraft_model=acmodelclass, flight_phase='reserve_cruise'),promotes_inputs=['ac|*'])
             # set the reserve_cruise time such that the desired design range is flown by the end of the mission
             phase5.add_subsystem('reserve_cruisedt',BalanceComp(name='duration',units='s',eq_units='m',val=120, upper=25000, lower=0,rhs_name='reserve_range',lhs_name='range_final'),promotes_outputs=['duration'])
             self.connect('reserve_range', 'reserve_cruise.reserve_cruisedt.reserve_range')
 
-            phase6 = self.add_subsystem('reserve_descent',NewSteadyFlightPhase(num_nodes=nn, aircraft_model=acmodelclass, flight_phase='reserve_descent'),promotes_inputs=['ac|*'])
+            phase6 = self.add_subsystem('reserve_descent',SteadyFlightPhase(num_nodes=nn, aircraft_model=acmodelclass, flight_phase='reserve_descent'),promotes_inputs=['ac|*'])
             # set the reserve_descent time so that the final altitude is sea level again
             phase6.add_subsystem('reserve_descentdt',BalanceComp(name='duration',units='s',eq_units='m', val=120, upper=8000, lower=0,rhs_name='takeoff|h',lhs_name='fltcond|h_final'),promotes_outputs=['duration'])
             phase6.connect('ode_integ.fltcond|h_final','reserve_descentdt.fltcond|h_final')
@@ -211,7 +211,7 @@ class MissionWithReserve(oc.TrajectoryGroup):
             self.connect('resrange.reserverange','reserve_cruise.reserve_cruisedt.range_final')
             # self.connect('reserve_descent.range_final', 'reserve_cruisedt.range_final')
 
-            phase7 = self.add_subsystem('loiter',NewSteadyFlightPhase(num_nodes=nn, aircraft_model=acmodelclass, flight_phase='loiter'),promotes_inputs=['ac|*'])
+            phase7 = self.add_subsystem('loiter',SteadyFlightPhase(num_nodes=nn, aircraft_model=acmodelclass, flight_phase='loiter'),promotes_inputs=['ac|*'])
             dvlist = [['duration_in', 'duration', 300, 's']]
             phase7.add_subsystem('loiter_dt', DVLabel(dvlist), promotes_inputs=["*"], promotes_outputs=["*"])
             self.connect('loiter|h0','loiter.ode_integ.fltcond|h_initial')
@@ -326,18 +326,18 @@ class FullMissionAnalysis(oc.TrajectoryGroup):
             self.add_subsystem('engineoutclimb',ClimbAnglePhase(num_nodes=1, aircraft_model=acmodelclass, flight_phase='EngineOutClimbAngle'), promotes_inputs=['ac|*'])
 
             # add the climb, cruise, and descent segments
-            climb = self.add_subsystem('climb',NewSteadyFlightPhase(num_nodes=nn, aircraft_model=acmodelclass, flight_phase='climb'),promotes_inputs=['ac|*'])
+            climb = self.add_subsystem('climb',SteadyFlightPhase(num_nodes=nn, aircraft_model=acmodelclass, flight_phase='climb'),promotes_inputs=['ac|*'])
             # set the climb time such that the specified initial cruise altitude is exactly reached
             climb.add_subsystem('climbdt',BalanceComp(name='duration',units='s',eq_units='m',val=120,lower=0,upper=3000,rhs_name='cruise|h0',lhs_name='fltcond|h_final'), promotes_outputs=['duration'])
             climb.connect('ode_integ.fltcond|h_final','climbdt.fltcond|h_final')
             self.connect('cruise|h0', 'climb.climbdt.cruise|h0')
 
-            cruise = self.add_subsystem('cruise',NewSteadyFlightPhase(num_nodes=nn, aircraft_model=acmodelclass, flight_phase='cruise'),promotes_inputs=['ac|*'])
+            cruise = self.add_subsystem('cruise',SteadyFlightPhase(num_nodes=nn, aircraft_model=acmodelclass, flight_phase='cruise'),promotes_inputs=['ac|*'])
             # set the cruise time such that the desired design range is flown by the end of the mission
             cruise.add_subsystem('cruisedt',BalanceComp(name='duration',units='s',eq_units='km',val=120, lower=0,upper=30000,rhs_name='mission_range',lhs_name='range_final'),promotes_outputs=['duration'])
             self.connect('mission_range', 'cruise.cruisedt.mission_range')
 
-            descent = self.add_subsystem('descent',NewSteadyFlightPhase(num_nodes=nn, aircraft_model=acmodelclass, flight_phase='descent'),promotes_inputs=['ac|*'])
+            descent = self.add_subsystem('descent',SteadyFlightPhase(num_nodes=nn, aircraft_model=acmodelclass, flight_phase='descent'),promotes_inputs=['ac|*'])
             # set the descent time so that the final altitude is sea level again
             descent.add_subsystem('descentdt',BalanceComp(name='duration',units='s',eq_units='m', val=120, lower=0,upper=3000,rhs_name='takeoff|h',lhs_name='fltcond|h_final'),promotes_outputs=['duration'])
             self.connect('takeoff|h','descent.descentdt.takeoff|h')
